Improving the ability to debug a circuit.
With that aim, we added the following three tools:

stoke_which_handler: Tells the name of the instruction handler responsible to
build the circuit for a given instruction. Usage: ./stoke_which_handler --opc
adcq_r64_r64

stoke_support_list: Prints the list of instructions supported by the current
stoke repo. Usage: ./stoke_suport_list

stoke_debug_formula: The augmented tool, similar to stoke_debug_formula, to
create  bit-vector formula for all the supported instructions. This
includes all the changes required for generating a formula for (1 )memory
instructions and (2) immediate instructions with artity 3, which is not
supported in the current Stoke branch. The memory is modelled using the trivial
memory model. Also, we added some switches in the tool to support the
following:

  * --keep_imm_symbolic: For an immediate instruction, the stoke_debug_formula
   application produces formulas for a particular instance of that
   instruction, not the generalized version. This switch will make sure that
   the immediate operand in the instruction is not instantiated to a
   particular value, but remain symbolically.

  // Before
  $ ./bin/stoke_debug_circuit --opc adcq_rax_imm32 (or use  --code "adcq \$0, %rax")
  %rax   : ((%cf ? 0x0₁ ∘ 0x0₆₄ + 0x1₆₅ : 0x0₁ ∘ 0x0₆₄) + 0x0₁ ∘ %rax)[63:0]  // This is the output state rax and the input immediate operand is chosen as 0.

  // After
  $ ./bin/stoke_debug_circuit --opc adcq_rax_imm32 --keep_imm_symbolic (or use  --code "adcq \$0, %rax")
  %rax   : ((%cf ? 0x0₁ ∘ sign-extend-64(Imm32) + 0x1₆₅ : 0x0₁ ∘ sign-extend-64(Imm32)) + 0x0₁ ∘ %rax)[63:0] // The input immediate operadn value is ignored. Used a symbolic variable instead.

  * --smtlib_formula: To print the Z3 formula corresponding to the output
   states of an instruction.

  * --opc: The tool can consume the opcode, like adcq_r64_r64,  over and above
   --code.