Return a valid pointer.
stringstream::str() creates a temporary that is destroyed at the end of the full-expression (the semi-colon), and c_str() on that string returns a pointer internal to the string. This function returns a pointer to freed memory.

Spotted by clang warning:
```
error.h:51:12: warning: returning address of local temporary object [-Wreturn-stack-address]
    return ss.str().c_str();
           ^~~~~~~~
1 warning generated.
```